requirements.txt
azure-functions
psycopg2-binary
httpx
google-auth
google-auth-oauthlib
google-api-python-client
google-genai
pytz
python-dotenv